+++ opensearchapi/api_reindex_rethrottle.go
@@ -21,7 +21,7 @@ func (c Client) ReindexRethrottle(ctx context.Context, req ReindexRethrottleReq)
 		data ReindexRethrottleResp
 		err  error
 	)
-	if data.response, err = c.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, &c, req, &data); err != nil {
 		return &data, err
 	}