+++ opensearchapi/api_update_by_query.go
@@ -22,7 +22,7 @@ func (c Client) UpdateByQuery(ctx context.Context, req UpdateByQueryReq) (*Updat
 		data UpdateByQueryResp
 		err  error
 	)
-	if data.response, err = c.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, &c, req, &data); err != nil {
 		return &data, err
 	}