+++ plugins/ism/api_refresh_search_analyzers.go
@@ -20,7 +20,7 @@ func (c Client) RefreshSearchAnalyzers(ctx context.Context, req RefreshSearchAna
 		data RefreshSearchAnalyzersResp
 		err  error
 	)
-	if data.response, err = c.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, &c, req, &data); err != nil {
 		return data, err
 	}