+++ plugins/ism/api_retry.go
@@ -23,7 +23,7 @@ func (c Client) Retry(ctx context.Context, req RetryReq) (RetryResp, error) {
 		data RetryResp
 		err  error
 	)
-	if data.response, err = c.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, &c, req, &data); err != nil {
 		return data, err
 	}