+++ plugins/security/api_actiongroups.go
@@ -24,7 +24,7 @@ func (c actiongroupsClient) Get(ctx context.Context, req *ActionGroupsGetReq) (A
 		data ActionGroupsGetResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data.Groups); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data.Groups); err != nil {
 		return data, err
 	}
 
@@ -37,7 +37,7 @@ func (c actiongroupsClient) Put(ctx context.Context, req ActionGroupsPutReq) (Ac
 		data ActionGroupsPutResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return data, err
 	}
 
@@ -50,7 +50,7 @@ func (c actiongroupsClient) Delete(ctx context.Context, req ActionGroupsDeleteRe
 		data ActionGroupsDeleteResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return data, err
 	}
 
@@ -63,7 +63,7 @@ func (c actiongroupsClient) Patch(ctx context.Context, req ActionGroupsPatchReq)
 		data ActionGroupsPatchResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return data, err
 	}