+++ plugins/security/api_securityconfig.go
@@ -24,7 +24,7 @@ func (c securityconfigClient) Get(ctx context.Context, req *ConfigGetReq) (Confi
 		data ConfigGetResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return data, err
 	}
 
@@ -37,7 +37,7 @@ func (c securityconfigClient) Put(ctx context.Context, req ConfigPutReq) (Config
 		data ConfigPutResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return data, err
 	}
 
@@ -50,7 +50,7 @@ func (c securityconfigClient) Patch(ctx context.Context, req ConfigPatchReq) (Co
 		data ConfigPatchResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return data, err
 	}